+++ lib/versions/onboarding_screen.dart
@@ -1,4 +1,4 @@
-import 'package:ble_uart/screens/patient_ver/register_screen.dart';
+import 'package:ble_uart/versions//patient_ver/register_screen.dart';
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_onboarding_slider/flutter_onboarding_slider.dart';
