+++ lib/versions/patient_ver/ai_screen.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'dart:convert';
-import 'package:ble_uart/screens/patient_ver/between_screen.dart';
-import 'package:ble_uart/screens/patient_ver/bottom_navigation_screen.dart';
+import 'package:ble_uart/versions//patient_ver/between_screen.dart';
+import 'package:ble_uart/versions//patient_ver/bottom_navigation_screen.dart';
 import 'package:ble_uart/utils/extra.dart';
 import 'package:ble_uart/utils/parsing_agc.dart';
 import 'package:flutter/cupertino.dart';
