+++ lib/versions/patient_ver/staff_screen.dart
@@ -1,7 +1,7 @@
-import 'package:ble_uart/screens/patient_ver/ai_screen.dart';
-import 'package:ble_uart/screens/patient_ver/database_screen.dart';
-import 'package:ble_uart/screens/patient_ver/threshold_setting_screen.dart';
-import 'package:ble_uart/screens/patient_ver/uart_screen.dart';
+import 'package:ble_uart/versions//patient_ver/ai_screen.dart';
+import 'package:ble_uart/versions//patient_ver/database_screen.dart';
+import 'package:ble_uart/versions//patient_ver/threshold_setting_screen.dart';
+import 'package:ble_uart/versions//patient_ver/uart_screen.dart';
 import 'package:ble_uart/utils/shared_prefs_utils.dart';
 import 'package:ble_uart/widgets/account_settings_tile.dart';
 import 'package:flutter/material.dart';
