+++ lib/versions/patient_ver/training_screen.dart
