+++ lib/widgets/register_form_field.dart
@@ -1,4 +1,4 @@
-import 'package:ble_uart/screens/patient_ver/first_connect_screen.dart';
+import 'package:ble_uart/versions//patient_ver/first_connect_screen.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
